fixing orb level up problem
